requirements.txt
rich
grpcio
grpcio-tools